feat: Enhance comment navigation - Allow navigation to previous/next comments without hierarchy
Here is the new behavior:

`h` and `l` => maps to next and previous comment without considering any hierarchy. Use this to navigate up and down every comment.
`j` and `k` => maps to next and previous comment at same depth. BUT, just like j which finds the next comment if there are no comments in the same depth, k also jumps to the previous comment if there are no comments at the same depth.

Resolves #3